Additional changes to parser. Separated classes into java and kotlin packages
