+++ codegen/src/main/java/org/web3j/codegen/unit/gen/java/JavaClassGenerator.java
@@ -10,20 +10,18 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.codegen.unit.gen;
+package org.web3j.codegen.unit.gen.java;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import javax.lang.model.element.Modifier;
 
 import com.squareup.javapoet.AnnotationSpec;
 import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.JavaFile;
-import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.TypeSpec;
 
+import org.web3j.codegen.unit.gen.UnitClassGenerator;
 import org.web3j.commons.JavaVersion;
 
 import static org.web3j.codegen.unit.gen.utills.NameUtils.toCamelCase;
