+++ codegen/src/main/java/org/web3j/codegen/unit/gen/java/MethodFilter.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.web3j.codegen.unit.gen;
+package org.web3j.codegen.unit.gen.java;
 
 import java.lang.reflect.Method;
 import java.util.ArrayList;
@@ -20,6 +20,8 @@ import java.util.stream.Collectors;
 
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.kotlinpoet.FunSpec;
+
+import org.web3j.codegen.unit.gen.kotlin.FunParser;
 import org.web3j.protocol.Web3j;
 import org.web3j.tx.TransactionManager;
 import org.web3j.tx.gas.ContractGasProvider;
@@ -43,7 +45,7 @@ public class MethodFilter {
             return Arrays.asList(method.getParameterTypes()).contains(Web3j.class)
                     && Arrays.asList(method.getParameterTypes()).contains(TransactionManager.class)
                     && Arrays.asList(method.getParameterTypes())
-                    .contains(ContractGasProvider.class);
+                            .contains(ContractGasProvider.class);
         }
         return true;
     }
