+++ codegen/src/test/java/org/web3j/codegen/unit/gen/MethodFilterTest.java
@@ -19,6 +19,7 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 
+import org.web3j.codegen.unit.gen.java.MethodFilter;
 import org.web3j.protocol.Web3j;
 import org.web3j.tx.TransactionManager;
 import org.web3j.tx.gas.ContractGasProvider;
